update trip details so we can use the new classnames
