+++ package.json
@@ -59,7 +59,7 @@
     "@opentripplanner/stops-overlay": "7.0.2",
     "@opentripplanner/transit-vehicle-overlay": "6.0.2",
     "@opentripplanner/transitive-overlay": "6.0.0",
-    "@opentripplanner/trip-details": "7.0.0",
+    "@opentripplanner/trip-details": "^8.0.1",
     "@opentripplanner/trip-form": "6.0.1",
     "@opentripplanner/trip-viewer-overlay": "4.0.1",
     "@opentripplanner/vehicle-rental-overlay": "4.0.2",
@@ -247,5 +247,6 @@
     "snapshot": {
       "percy-css": ".percy-hide { opacity: 0!important; } "
     }
-  }
+  },
+  "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e"
 }
